API

Table of Contents

QGIS.df

QGIS.Algorithm

QGIS.help

QGIS.metadata

Docstrings

QGIS.df — Constant

QGIS.df

A DataFrame containing the metadata for all available QGIS algorithms.

QGIS.Algorithm — Type

QGIS.Algorithm(name)

Object representing a QGIS algorithm. The struct contains the help, metadata, and default parameters for the algorithm. Algorithms are callable and expect a String input (filename) and optional String output (filename).

Example

using QGIS, GeoJSON, Plots, GeoInterfaceRecipes

buffer = QGIS.Algorithm("native:buffer", DISTANCE=0.1)

input = joinpath(dirname(pathof(QGIS)), "..", "test", "nc.geojson")

nc = GeoJSON.read(input)

output = buffer(input)

nc_buffered = GeoJSON.read(output)

plot(
    plot(nc.geometry),
    plot(nc_buffered.geometry),
    layout = (2, 1)
)

QGIS.help — Method

help(alg)

Return the help (in JSON format) for the associated algorithm.

QGIS.metadata — Method

metadata(alg)

Return the associated metadata for a given algorithm. Returns the associated DataFrameRow from QGIS.df.